% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsPlotR_library.R
\name{bb.calcTSNE}
\alias{bb.calcTSNE}
\title{Calculates the bhSNE from a melted data table}
\usage{
bb.calcTSNE(input_dat, channels, value_var = "counts",
  channel_var = "channel", id_var = "id", group_var = "condition",
  scale = F, subsample_groups = F, subsample_mode = "equal",
  verbose = T, dims = 2, ...)
}
\arguments{
\item{input_dat}{melted data_table}

\item{channels}{list of channels to be Used}

\item{value_var}{column to be used as values}

\item{channel_var}{column with the channel names}

\item{id_var}{column name for the unique cell ID}

\item{group_var}{variable that groups the table}

\item{scale}{should the input data be rescaled?}

\item{subsample_groups}{T: resamples all groups to the group with the least members, Integer: resamples all groups to the integer}

\item{subsample_mode}{'equal': if integer provided to resample_groups all groups will contain min(resample_groups, min_groupsize) cells
'unequal': if integer provided to resample_groups, all groups will contain maximally resample_groups cells (or less)}

\item{verbose:}{should bhSne be verbose?}
}
\value{
tsne object
}
\description{
Calculates the bhSNE from a melted data table
}

