% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsPlotR_library.R
\name{bb.flowFrame2dt}
\alias{bb.flowFrame2dt}
\title{Converts FlowFrame to data table}
\usage{
bb.flowFrame2dt(datFCS)
}
\arguments{
\item{datFCS}{A flowframe e.g. from read.FCS or bb.loadFCS}
}
\value{
A data.table with row=cell, column=Channel
}
\description{
Converts FlowFrame to data table
}
\examples{
read.FCS('/test/testfcs.fcs')
}

