% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsPlotR_library.R
\name{bb.getInfoFromFileList}
\alias{bb.getInfoFromFileList}
\title{Parses a list of strings and gets the corresponding information}
\usage{
bb.getInfoFromFileList(fileList, sep = "_", strPos = 2,
  censorStr = ".fcs")
}
\description{
See bb.getInfoFromString for more information
}

