% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsPlotR_library.R
\name{bb.loadConvertMultiFCS}
\alias{bb.loadConvertMultiFCS}
\title{Loads multiple FCS files at once}
\usage{
bb.loadConvertMultiFCS(fileList = NaN, fileDir = NaN, condDict = NaN,
  subSample = NA)
}
\arguments{
\item{fileList}{List of files to be loaded}

\item{fileDir}{the base directory for the files or a directory with FCS files}

\item{condDict}{a named vector which maps Filenames to condition}
}
\value{
a dataframe with all combined FCS files
}
\description{
Loads multiple FCS files at once
}

