% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsPlotR_library.R
\name{bb.loadFCS}
\alias{bb.loadFCS}
\title{Wrapper for flow Core read.FCS}
\usage{
bb.loadFCS(filePath, ...)
}
\arguments{
\item{filePath}{path to the FCS file}
}
\value{
A flowframe object
}
\description{
Wrapper for flow Core read.FCS
}
\examples{
read.FCS('/test/testfcs.fcs')
}

