% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{censor_dat}
\alias{censor_dat}
\title{censor dat
remove the outliers on the upper side by capping the values at the provided quantile}
\usage{
censor_dat(x, quant = 0.999, symmetric = F)
}
\arguments{
\item{x}{values to censor}

\item{quant}{quantile to censor, i.e. how many percent of values are considered outliers}

\item{symmetric}{censor on both side. In this case the outliers are assumed to be symetric on both sides. For example if a quantile of 5\% (0.05) is choosen, in the symetric case 2.5\% (0.025) of values are censored on both sides.}
}
\value{
returns the percentile of each value of x
}
\description{
censor dat
remove the outliers on the upper side by capping the values at the provided quantile
}
