% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsPlotR_library.R
\name{do_flowsom}
\alias{do_flowsom}
\title{Run flowsom on a melted data table}
\usage{
do_flowsom(data, channels, valuevar = "counts_transf",
  channelvar = "channel", idvar = "id", k = 20, seed = FALSE,
  subsample = FALSE, return_output = FALSE)
}
\arguments{
\item{data}{a data frame in the long format}

\item{channels}{a list of channel names to use}

\item{valuevar}{the column to take as value variable}

\item{channelvar}{the column to take as channel variable}

\item{idvar}{the column to toake as id variable}

\item{k}{the number of clusters to use for metaclustering}

\item{seed}{the random seed - if provided}

\item{subsample}{number of cells to use for subsampling - if provided}

\item{return_output}{should the cluster algorithm output be provided? If so the return value will be a list with the table and the original output.}
}
\value{
at table with a column cluster and a column idvar
}
\description{
Run flowsom on a melted data table
}
