% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsPlotR_library.R
\name{do_phenograph}
\alias{do_phenograph}
\title{Make a phenograph from a melted dataframe}
\usage{
do_phenograph(data, channels, valuevar = "counts_transf",
  channelvar = "channel", idvar = "id", k = 20, seed = FALSE,
  subsample = FALSE)
}
\arguments{
\item{data}{a data frame in the long format}

\item{channels}{a list of channel names to use}

\item{valuevar}{the column to take as value variable}

\item{channelvar}{the column to take as channel variable}

\item{idvar}{the column to toake as id variable}

\item{k}{the number of nearest neighbours}

\item{seed}{the random seed - if provided}

\item{subsample}{number of cells to use for subsampling - if provided}
}
\value{
at table with a column cluster and a column idvar
}
\description{
Make a phenograph from a melted dataframe
}
