% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsPlotR_library.R
\name{get_cormat}
\alias{get_cormat}
\title{Calculate a correlation matrix from a data table}
\usage{
get_cormat(data, xcol, ycol, valuecol, method = "pearson", pval = F)
}
\arguments{
\item{data:}{a data table}

\item{xcol:}{the column name to be taken as x column}

\item{ycol:}{the column name to be taken as y column}

\item{valuecol:}{the column name to be taken as valuecol}

\item{method:}{pearson or spearman, type argumetn from 'rcorr'}

\item{pval:}{return p values for the correlations?}
}
\description{
Calculate a correlation matrix from a data table
}
