% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{getInfoFromFileList}
\alias{getInfoFromFileList}
\title{Parses a list of strings and gets the corresponding information}
\usage{
getInfoFromFileList(fileList, sep = "_", strPos = 2, censorStr = ".fcs")
}
\description{
See getInfoFromString for more information
}
