% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{getInfoFromString}
\alias{getInfoFromString}
\title{Generates a metadata table from a}
\usage{
getInfoFromString(name, sep = "_", strPos = 2, censorStr = ".fcs")
}
\arguments{
\item{name}{a string}

\item{strPos}{numeric or vector: indicates which positions of the string should be extracted}

\item{censorStr}{string that gets ignored}

\item{seq}{the seperator that seperates fields in the string}
}
\value{
The extracted stringfield
}
\description{
Extracts information from string fields
}
