% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{getPerc}
\alias{getPerc}
\title{calculate percentiles
For each value in x  calculate which percentile it is}
\usage{
getPerc(x)
}
\arguments{
\item{x}{values to be mapped}
}
\value{
returns the percentile of each value of x
}
\description{
calculate percentiles
For each value in x  calculate which percentile it is
}
