% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsPlotR_library.R
\name{load_json_graph}
\alias{load_json_graph}
\title{Loads a Cytoscape JSON as an igraph graph}
\usage{
load_json_graph(jdat)
}
\arguments{
\item{jdat}{a loaded Cytoscape JSON}
}
\value{
an igraph-graph
}
\description{
Loads a Cytoscape JSON as an igraph graph
}
\examples{

\dontrun{
jdat = fromJSON(file = 'xy.json')
g = load_json_graph(jdat)
list.edge.attributes(g)
list.vertex.attributes(g)

plot.igraph(g,
           layout=-cbind(V(g)$posx, V(g)$posy),
           rescale=T,
           edge.arrow.size=0.1,
           vertex.size =7,
           vertex.label.cex = 0.5)
}
}
