% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsPlotR_library.R
\name{loadConvertMultiFCS}
\alias{loadConvertMultiFCS}
\title{Loads multiple FCS files at once}
\usage{
loadConvertMultiFCS(fileList = NaN, fileDir = NaN, condDict = NaN,
  subSample = NA)
}
\arguments{
\item{fileList}{Vector of filename to be loaded. If NaN all fcs files are loaded}

\item{fileDir}{the base directory for the files or a directory with FCS files}

\item{condDict}{a named vector which maps Filenames to condition}
}
\value{
a data/tab;e with all combined FCS files

This is a good way to load a whole directory of files (if fileList=NaN).
If a condition dictionary (condDict) is provided, the 'condition' column will
be equal to the conditions. Otherwise it will be the filename used.
}
\description{
Loads multiple FCS files at once
}
