% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{loadFCS}
\alias{loadFCS}
\title{Wrapper for flow Core read.FCS}
\usage{
loadFCS(filePath, ...)
}
\arguments{
\item{filePath}{path to the FCS file}
}
\value{
A flowframe object
}
\description{
Wrapper for flow Core read.FCS
}
