% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsPlotR_library.R
\name{map2colormap}
\alias{map2colormap}
\title{Maps a vector x to a colormap and returns the RGB values}
\usage{
map2colormap(x, xmax = NA, symmetric = NA, cmap = colorRamp(c("blue",
  "white", "red")), na_col = "cornsilk3")
}
\arguments{
\item{x}{values to be mapped}

\item{xmax}{maximum value to be mapped}

\item{symmetric}{if T the map will be symmetric from -xmax to xmax centered arround 0}

\item{cmap}{a colormap, e.g. from colorRamp}

\item{na_col}{what color should NA have (name or RGB)}
}
\value{
returns the values mapped
}
\description{
Maps a vector x to a colormap and returns the RGB values
}
