% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{plot_sumStats}
\alias{plot_sumStats}
\title{plot summary stats}
\usage{
plot_sumStats(df, varName = "value", condName = "condition",
  channelName = "channel", fkt = function(x) {     x })
}
\description{
plot summary stats
}
