% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{censor_dat}
\alias{censor_dat}
\title{censor dat
remove the outliers on the upper side by capping the values at the provided quantile}
\usage{
censor_dat(x, quant = 0.999, symmetric = F)
}
\arguments{
\item{x}{values to censor}

\item{quant}{quantile to censor by}

\item{symmetric}{censor on both side. The quantile is then seen as the inner quantile.}
}
\value{
returns the percentile of each value of x
}
\description{
censor dat
remove the outliers on the upper side by capping the values at the provided quantile
}
